+++ document/src/epilot/models/operations/generatedocument.py
@@ -64,9 +64,12 @@ class GenerateDocument200ApplicationJSON:
 @dataclasses.dataclass
 class GenerateDocumentResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     generate_document_200_application_json_object: Optional[GenerateDocument200ApplicationJSON] = dataclasses.field(default=None)
     r"""Generated document output"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""