+++ document/src/epilot/models/operations/generatedocumentv2.py
@@ -161,9 +161,12 @@ class GenerateDocumentV2200ApplicationJSON:
 @dataclasses.dataclass
 class GenerateDocumentV2Response:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     generate_document_v2_200_application_json_object: Optional[GenerateDocumentV2200ApplicationJSON] = dataclasses.field(default=None)
     r"""Generated document output"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""